% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{get_models_dir}
\alias{get_models_dir}
\title{Get the local models directory}
\usage{
get_models_dir()
}
\value{
Character string path to local models directory
}
\description{
Returns the path to the local models cache directory. By default uses
the inst/models directory within the package, but can be overridden
via the torpmodels.cache_dir option.
}
\keyword{internal}
