% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{get_torpmodels_repo}
\alias{get_torpmodels_repo}
\title{Get the torpmodels repository}
\usage{
get_torpmodels_repo()
}
\value{
Character string of the repository in format "owner/repo"
}
\description{
Get the torpmodels repository
}
\keyword{internal}
