% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{list_available_models}
\alias{list_available_models}
\title{List Available Models}
\usage{
list_available_models()
}
\value{
A list with two elements: core_models and stat_models
}
\description{
Returns a list of available models that can be loaded.
}
\examples{
list_available_models()
}
