% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{normalize_model_name}
\alias{normalize_model_name}
\title{Normalize model name to file and tag}
\usage{
normalize_model_name(model_name)
}
\description{
Normalize model name to file and tag
}
\keyword{internal}
