% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{download_model_from_release}
\alias{download_model_from_release}
\title{Download model from GitHub release}
\usage{
download_model_from_release(file_name, release_tag, local_path, verbose = TRUE)
}
\description{
Download model from GitHub release
}
\keyword{internal}
